adding curd operation on employees

+++ src/main/java/com/preet/repository/EmployeeDAO.java
@@ -3,6 +3,6 @@ package com.preet.repository;
 import com.preet.entity.EmployeeEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmployeeDAO extends JpaRepository<EmployeeEntity, Long> {
+public interface EmployeeDAO extends JpaRepository<EmployeeEntity, Integer> {
 
 }
